renamed coordinates component to pages
