+++ src/components/coordinatetransformation/InputCard.vue
@@ -4,7 +4,7 @@
       <h3>Input</h3>
     </div>
     <section class="coordinate-selection-wrapper">
-      <CoordinateSelection :isOutput="false" @epsg-changed="inputEPSGChanged"/>
+      <EpsgSelection :isOutput="false" @epsg-changed="inputEPSGChanged"/>
     </section>
     <div class="input">
       <span class="first-input" :class="{isDegreesInput: isDegrees, isMetresInput: !isDegrees}">
@@ -208,10 +208,10 @@ import { defineAsyncComponent, ref, inject, onUpdated, watch, onMounted } from '
 import { useStore } from 'vuex'
 
 export default {
-  name: 'InputCoordinates',
+  name: 'InputCardComponent',
 
   components: {
-    CoordinateSelection: defineAsyncComponent(() => import('@/components/coordinatetransformation/CoordinateSelection'))
+    EpsgSelection: defineAsyncComponent(() => import('@/components/coordinatetransformation/EpsgSelection'))
   },
 
   methods: {
